requirements.txt
snntorch
transformers
datasets
accelerate
wandb
pyyaml